+++ packages/generator-adp/src/app/index.ts
@@ -1,26 +1,28 @@
+import { join } from 'path';
 import Generator from 'yeoman-generator';
 import { AppWizard, Prompts } from '@sap-devx/yeoman-ui-types';
 
-import { ToolsLogger } from '@sap-ux/logger';
-import type { ConfigAnswers, FlexLayer } from '@sap-ux/adp-tooling';
-import { isInternalFeaturesSettingEnabled } from '@sap-ux/feature-toggle';
-import { SystemLookup, generate, getConfig } from '@sap-ux/adp-tooling';
 import {
     TelemetryHelper,
     sendTelemetry,
     type ILogWrapper,
     getHostEnvironment,
-    hostEnvironment
+    hostEnvironment,
+    getDefaultTargetFolder
 } from '@sap-ux/fiori-generator-shared';
+import { ToolsLogger } from '@sap-ux/logger';
+import { SystemLookup, generate, getConfig } from '@sap-ux/adp-tooling';
+import { isInternalFeaturesSettingEnabled } from '@sap-ux/feature-toggle';
+import type { AttributesAnswers, ConfigAnswers, FlexLayer } from '@sap-ux/adp-tooling';
 
 import { getFlexLayer } from './layer';
 import { t, initI18n } from '../utils/i18n';
 import { EventName } from '../telemetryEvents';
 import AdpFlpConfigLogger from '../utils/logger';
-import type { AdpGeneratorOptions } from './types';
+import { getPrompts } from './questions/attributes';
 import { ConfigPrompter } from './questions/configuration';
 import { getPackageInfo, installDependencies } from '../utils/deps';
-import { generateValidNamespace, getDefaultProjectName } from './questions/helper/default-values';
+import type { AdpGeneratorOptions, AttributePromptOptions } from './types';
 
 /**
  * Generator for creating an Adaptation Project.
@@ -54,9 +56,9 @@ export default class extends Generator {
      */
     private configAnswers: ConfigAnswers;
     /**
-     * Target folder for the generated project.
+     * Project attribute answers.
      */
-    private targetFolder: string;
+    private attributeAnswers: AttributesAnswers;
     /**
      * SystemLookup instance for managing system endpoints.
      */
@@ -87,12 +89,16 @@ export default class extends Generator {
     async initializing(): Promise<void> {
         await initI18n();
 
-        const pages = [{ name: t('yuiNavSteps.configurationName'), description: t('yuiNavSteps.configurationDescr') }];
+        const pages = [
+            { name: t('yuiNavSteps.configurationName'), description: t('yuiNavSteps.configurationDescr') },
+            { name: t('yuiNavSteps.projectAttributesName'), description: t('yuiNavSteps.projectAttributesDescr') }
+        ];
         this.prompts.splice(0, 0, pages);
 
         this.layer = await getFlexLayer();
 
         this.systemLookup = new SystemLookup(this.toolsLogger);
+        this.prompter = new ConfigPrompter(this.systemLookup, this.layer, this.toolsLogger);
 
         await TelemetryHelper.initTelemetrySettings({
             consumerModule: {
@@ -105,7 +111,6 @@ export default class extends Generator {
     }
 
     async prompting(): Promise<void> {
-        this.prompter = new ConfigPrompter(this.systemLookup, this.layer, this.toolsLogger);
         const isCLI = getHostEnvironment() === hostEnvironment.cli;
 
         const configQuestions = this.prompter.getPrompts({
@@ -117,25 +122,40 @@ export default class extends Generator {
 
         this.logger.info(`System: ${this.configAnswers.system}`);
         this.logger.info(`Application: ${JSON.stringify(this.configAnswers.application, null, 2)}`);
+
+        const { ui5Versions, systemVersion } = this.prompter.ui5;
+        const promptConfig = {
+            ui5Versions,
+            isVersionDetected: !!systemVersion,
+            isCloudProject: this.prompter.isCloud,
+            layer: this.layer
+        };
+        const defaultFolder = getDefaultTargetFolder(this.options.vscode) ?? process.cwd();
+        const options: AttributePromptOptions = {
+            targetFolder: { default: defaultFolder },
+            ui5ValidationCli: { hide: !isCLI }
+        };
+        const attributesQuestions = getPrompts(this.destinationPath(), promptConfig, options);
+
+        this.attributeAnswers = await this.prompt(attributesQuestions);
+
+        this.logger.info(`Project Attributes: ${JSON.stringify(this.attributeAnswers, null, 2)}`);
     }
 
     async writing(): Promise<void> {
         try {
-            const projectName = getDefaultProjectName(this.destinationPath());
-            const namespace = generateValidNamespace(projectName, this.layer);
-            this.targetFolder = this.destinationPath(projectName);
-
             const packageJson = getPackageInfo();
             const config = await getConfig({
                 provider: this.prompter.provider,
                 configAnswers: this.configAnswers,
+                attributeAnswers: this.attributeAnswers,
+                publicVersions: this.prompter?.ui5?.publicVersions,
                 layer: this.layer,
-                defaults: { namespace },
                 packageJson,
                 logger: this.toolsLogger
             });
 
-            await generate(this.targetFolder, config, this.fs);
+            await generate(this._getProjectPath(), config, this.fs);
         } catch (e) {
             this.logger.error(`Writing phase failed: ${e}`);
             throw new Error(t('error.updatingApp'));
@@ -145,7 +165,7 @@ export default class extends Generator {
     async install(): Promise<void> {
         try {
             if (this.shouldInstallDeps) {
-                await installDependencies(this.targetFolder);
+                await installDependencies(this._getProjectPath());
             }
         } catch (e) {
             this.logger.error(`Installation of dependencies failed: ${e.message}`);
@@ -159,10 +179,21 @@ export default class extends Generator {
                 ...this.options.telemetryData
             }) ?? {};
         if (telemetryData) {
-            sendTelemetry(EventName.ADAPTATION_PROJECT_CREATED, telemetryData, this.targetFolder).catch((error) => {
-                this.logger.error(t('error.telemetry', { error }));
-            });
+            sendTelemetry(EventName.ADAPTATION_PROJECT_CREATED, telemetryData, this._getProjectPath()).catch(
+                (error) => {
+                    this.logger.error(t('error.telemetry', { error }));
+                }
+            );
         }
+    }
+
+    /**
+     * Combines the target folder and project name.
+     *
+     * @returns {string} The project path from the answers.
+     */
+    private _getProjectPath(): string {
+        return join(this.attributeAnswers.targetFolder, this.attributeAnswers.projectName);
     }
 
     /**