+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.1.47
+
+### Patch Changes
+
+-   Updated dependencies [ce3fd7a]
+    -   @sap-ux/cf-deploy-config-writer@0.2.4
+
+## 0.1.46
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.36
+-   @sap-ux/cf-deploy-config-inquirer@0.2.29
+-   @sap-ux/deploy-config-generator-shared@0.0.42
+
 ## 0.1.45
 
 ### Patch Changes