+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 0.1.12
+
+### Patch Changes
+
+-   @sap-ux/cf-deploy-config-sub-generator@0.1.47
+
+## 0.1.11
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.36
+-   @sap-ux/odata-service-inquirer@2.2.36
+-   @sap-ux/abap-deploy-config-sub-generator@0.0.62
+-   @sap-ux/cf-deploy-config-sub-generator@0.1.46
+-   @sap-ux/deploy-config-generator-shared@0.0.42
+
+## 0.1.10
+
+### Patch Changes
+
+-   @sap-ux/odata-service-inquirer@2.2.35
+
+## 0.1.9
+
+### Patch Changes
+
+-   Updated dependencies [1bc1cf2]
+    -   @sap-ux/odata-service-inquirer@2.2.34
+
 ## 0.1.8
 
 ### Patch Changes