+++ packages/generator-adp/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/generator-adp
 
+## 0.1.10
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.36
+-   @sap-ux/adp-tooling@0.13.31
+
+## 0.1.9
+
+### Patch Changes
+
+-   2db2c33: feat(generator-adp): Add Project Attributes page
+-   Updated dependencies [2db2c33]
+    -   @sap-ux/project-input-validator@0.5.1
+    -   @sap-ux/adp-tooling@0.13.30
+
 ## 0.1.8
 
 ### Patch Changes