+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/odata-service-inquirer
 
+## 2.2.36
+
+### Patch Changes
+
+-   Updated dependencies [e5f4915]
+    -   @sap-ux/guided-answers-helper@0.2.2
+    -   @sap-ux/inquirer-common@0.6.36
+
+## 2.2.35
+
+### Patch Changes
+
+-   Updated dependencies [2db2c33]
+    -   @sap-ux/project-input-validator@0.5.1
+
+## 2.2.34
+
+### Patch Changes
+
+-   1bc1cf2: Fix for v2 annotations not downloaded for generic destinations (full/partial urls)
+
 ## 2.2.33
 
 ### Patch Changes