+++ packages/ui-service-sub-generator/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/ui-service-sub-generator
 
+## 0.0.21
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.36
+-   @sap-ux/ui-service-inquirer@0.0.21
+
+## 0.0.20
+
+### Patch Changes
+
+-   @sap-ux/ui-service-inquirer@0.0.20
+
+## 0.0.19
+
+### Patch Changes
+
+-   @sap-ux/ui-service-inquirer@0.0.19
+
 ## 0.0.18
 
 ### Patch Changes